backend/requirements.txt
fastapi
psycopg2-binary
uvicorn
bcrypt
pyjwt
typing